add loading of past body weight entries grouped into time intervals

+++ app/src/main/java/com/dominikgold/calorietracker/entities/BodyWeightEntry.kt
@@ -8,5 +8,5 @@ import java.time.LocalDate
  */
 data class BodyWeightEntry(
     val date: LocalDate,
-    val bodyWeight: Float,
+    val bodyWeight: Double,
 )
